Readme.me
It is a web based application for managing the restaurant. This project mainly focuses on providing information to the user about restaurants in nearby places. The project is developed so such that users can view the information about restaurant which they want to visit and know food availability, ordering online and can also do booking of the table where we want and make online payment in just your hand. The project also has a login page where the user can basically get all the information about the place and can accordingly plan at which period the customer wants to visit and save his precious time. The main aim of developing this application is to reduce the time to a great extent that is spent in searching about where to go? Thus this application provides the required information in no time and also helps in quicker decision making through the mobiles on the hand.